+++ lib/User/favorite.dart
@@ -170,40 +170,41 @@ class PlaceList extends StatelessWidget {
                           context,
                           MaterialPageRoute(
                               builder: (context) => Business_Detail(
-                                    place_id: doc["place_id"],
-                                    address: doc["address"],
-                                    business_name: doc["business_name"],
-                                    business_name1: doc["business_name1"],
-                                    business_name2: doc["business_name2"],
-                                    business_name3: doc["business_name3"],
+                                    place_id: document["place_id"],
+                                    address: document["address"],
+                                    business_name: document["business_name"],
+                                    business_name1: document["business_name1"],
+                                    business_name2: document["business_name2"],
+                                    business_name3: document["business_name3"],
                                     business_name_english:
-                                        doc["business_name_english"],
-                                    day: doc["day"],
-                                    detail: doc["detail"],
-                                    email: doc["email"],
-                                    facebook: doc["facebook"],
-                                    instagram: doc["instagram"],
-                                    line: doc["line"],
-                                    latitude: doc["latitude"],
-                                    longitude: doc["longitude"],
-                                    map: doc["map"],
-                                    photo1: doc["photo1"],
-                                    photo2: doc["photo2"],
-                                    photo3: doc["photo3"],
-                                    photo4: doc["photo4"],
-                                    photo5: doc["photo5"],
-                                    photo6: doc["photo6"],
-                                    photo7: doc["photo7"],
-                                    photo8: doc["photo8"],
-                                    photo9: doc["photo9"],
-                                    photo10: doc["photo10"],
-                                    price: doc["price"],
-                                    rating: doc["rating"],
-                                    tel: doc["tel"],
-                                    time: doc["time"],
-                                    type: doc["type"],
-                                    user_id: doc["user_id"],
-                                    website: doc["website"],
+                                        document["business_name_english"],
+                                    day: document["day"],
+                                    detail: document["detail"],
+                                    email: document["email"],
+                                    facebook: document["facebook"],
+                                    instagram: document["instagram"],
+                                    line: document["line"],
+                                    latitude: document["latitude"],
+                                    longitude: document["longitude"],
+                                    map: document["map"],
+                                    photo1: document["photo1"],
+                                    photo2: document["photo2"],
+                                    photo3: document["photo3"],
+                                    photo4: document["photo4"],
+                                    photo5: document["photo5"],
+                                    photo6: document["photo6"],
+                                    photo7: document["photo7"],
+                                    photo8: document["photo8"],
+                                    photo9: document["photo9"],
+                                    photo10: document["photo10"],
+                                    price: document["price"],
+                                    rating:
+                                        document["rating"].toStringAsFixed(1),
+                                    tel: document["tel"],
+                                    time: document["time"],
+                                    type: document["type"],
+                                    user_id: document["user_id"],
+                                    website: document["website"],
                                   )),
                         );
                       },
