+++ pubspec.yaml
@@ -30,7 +30,6 @@ dependencies:
 
   cupertino_icons: ^1.0.3
   firebase_core: ^1.4.0
-  firebase_database: ^7.1.2
   cloud_firestore: ^2.4.0
   firebase_storage: ^10.0.1
   progress_dialog: ^1.2.4
